requirements.txt
aiobotocore==1.2.1
aiohttp==3.7.3
aioitertools==0.7.1
async-timeout==3.0.1
attrs==20.3.0
aws-cdk.assets==1.90.1
aws-cdk.aws-apigateway==1.90.1
aws-cdk.aws-apigatewayv2==1.90.1
aws-cdk.aws-applicationautoscaling==1.90.1
aws-cdk.aws-autoscaling==1.90.1
aws-cdk.aws-autoscaling-common==1.90.1
aws-cdk.aws-autoscaling-hooktargets==1.90.1
aws-cdk.aws-batch==1.90.1
aws-cdk.aws-certificatemanager==1.90.1
aws-cdk.aws-cloudformation==1.90.1
aws-cdk.aws-cloudfront==1.90.1
aws-cdk.aws-cloudwatch==1.90.1
aws-cdk.aws-codebuild==1.90.1
aws-cdk.aws-codecommit==1.90.1
aws-cdk.aws-codeguruprofiler==1.90.1
aws-cdk.aws-codepipeline==1.90.1
aws-cdk.aws-cognito==1.90.1
aws-cdk.aws-ec2==1.90.1
aws-cdk.aws-ecr==1.90.1
aws-cdk.aws-ecr-assets==1.90.1
aws-cdk.aws-ecs==1.90.1
aws-cdk.aws-efs==1.90.1
aws-cdk.aws-elasticloadbalancing==1.90.1
aws-cdk.aws-elasticloadbalancingv2==1.90.1
aws-cdk.aws-events==1.90.1
aws-cdk.aws-events-targets==1.90.1
aws-cdk.aws-iam==1.90.1
aws-cdk.aws-kinesis==1.90.1
aws-cdk.aws-kinesisfirehose==1.90.1
aws-cdk.aws-kms==1.90.1
aws-cdk.aws-lambda==1.90.1
aws-cdk.aws-lambda-python==1.90.1
aws-cdk.aws-logs==1.90.1
aws-cdk.aws-route53==1.90.1
aws-cdk.aws-route53-targets==1.90.1
aws-cdk.aws-s3==1.90.1
aws-cdk.aws-s3-assets==1.90.1
aws-cdk.aws-sam==1.90.1
aws-cdk.aws-secretsmanager==1.90.1
aws-cdk.aws-servicediscovery==1.90.1
aws-cdk.aws-sns==1.90.1
aws-cdk.aws-sns-subscriptions==1.90.1
aws-cdk.aws-sqs==1.90.1
aws-cdk.aws-ssm==1.90.1
aws-cdk.aws-stepfunctions==1.90.1
aws-cdk.cloud-assembly-schema==1.90.1
aws-cdk.core==1.90.1
aws-cdk.custom-resources==1.90.1
aws-cdk.cx-api==1.90.1
aws-cdk.region-info==1.90.1
beautifulsoup4==4.9.3
boto3==1.17.12
botocore==1.20.12
cattrs==1.1.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
constructs==3.3.29
Cython==0.29.21
Flask==1.1.2
Flask-Cors==3.0.10
fsspec==0.8.5
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsii==1.21.0
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.1
pandas==1.3.0.dev0+741.g4f14b24ff
pandasql==0.7.3
publication==0.0.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3fs==0.5.2
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
